refactor(event-details): decompose screen into hooks and components
Extract useEventDetailsData, useEventDetailsActions, and
useHostRequestActions hooks plus EventDetailsHero, EventDetailsInfo,
HostRequestTabs, EventDetailsMembers, and EventDetailsCTA components,
with styles in EventDetailsScreen.styles.ts using theme tokens where
exact matches exist. Screen drops from 2,197 to 582 lines. Behavior
unchanged; all 107 existing Event Details tests pass unmodified.
